Add TODOs for future work. I feel like the codebase of this file has been  outgrowing the inline-shell-single-gitconfig paradigm for a while, and as such will need to be refactored (at the very minimum) so that each git subcommand defined here is a separate shell script, or a call to a separate shell script with some parameters passed in.
The benefit of this will be the ability to reuse code cleanly, having
consistent autocomplete without having to force the use of pure-git-subcommand
aliases, maybe some help text, and definitely easier to understand concept
space of what some of the commands do if they are a shell-out to the same
script. Perhaps some of them might go away, since they are literally just
shell-outs with varied parameters and would be better off if they did
shell out, but to a better-defined script with a more human-suitable
conceptual space.

Side note: per-branch scripting/variable setting could be achieved
with a post-checkout hook (see https://stackoverflow.com/questions/49827353/using-gitconfig-per-branch)